requirements.txt
openai>=1.40.0
opentelemetry-api>=1.28.0
opentelemetry-sdk>=1.28.0
opentelemetry-instrumentation-openai-v2>=2.0.0
azure-monitor-opentelemetry-exporter
azure-ai-projects>=1.0.0
azure-identity>=1.19.0
python-dotenv>=1.0.0
